Expand "advanced" API docs
This PR increases the verbosity of the "advanced techniques" page of
the documentation. The intent to be more explicit about what
occurs with default behavior, and how customizations yield a different
outcome.

In some ways this is unfinished: the portion of these docs involving
`Location` may be a bug in StructArrays. I didn't touch this section
of the documentation.