Docs: improve copy-vs-view, mutability, and advanced API (#225)
* Switch docs from README to hosted Documenter

These were near-duplicates and risked duplicating effort.

* Document the `dims` constructor in the introduction

* Disentangle mutability from AOS-copying behavior

In the "counterintuitive" section, the documentation formerly mixed
together two concepts: mutable structs and the copying behavior when
constructing a StructArray from an Array-of-Structs. This commit
disentagles these concepts.

* Expand "advanced" API docs

Increases the verbosity of the "advanced techniques" page of
the documentation. The intent to be more explicit about what
occurs with default behavior, and how customizations yield a different
outcome.

* Disable doctesting on versions prior to Julia 1.6

Also adds Julia 1.6LTS to the CI matrix.

Co-authored-by: Jesse Chan <jchan985@gmail.com>