Convert to `StaticArrayStyle`
We first call broadcast from `StaticArrays` then split the output.
This should has no extra runtime overhead. But some type info might missing because the eltype change. I think there's no better ways as we don't want to depend on the full  `StaticArrays`.

We don't overloading `Size` and `similar_type` at present.
as they are only used for `broadcast`.
With this, we can move much less code to `StaticArraysCore`.

The only downside is that SizedArray would be allocated twice. That's not idea, but we can't do any better if we don't depend on StaticArray or copy a lot of code from there.