Fix incorrect computation of return_type of `itr`
The `itr` return type was assumed to be always the same as the first
element. This is false for many objects, such as Tuple. Additionally,
it was using a private API (Core.Compiler and
Core.Compiler.return_type) which are not stable across versions of
Julia and may return weird and unpredictable answers. Switch to using
the API in Base for this.